+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 - Rename `desc` in APIs to `doc` to avoid confusion with the `description` field of manpages. (#12)
 - Basic manpage generation (#13)
+- Command aliases (#14)
 
 ## 0.4.0